refactor: dedupe storybook connected draggable components
